Move uninhabited unreachable code lint to its own pass.
