+++ tests/mir-opt/building/issue_101867.main.built.after.mir
@@ -32,7 +32,7 @@ fn main() -> () {
     bb1: {
         StorageLive(_3);
         StorageLive(_4);
-        _4 = std::rt::begin_panic::<&str>(const "explicit panic") -> bb8;
+        _4 = std::rt::begin_panic::<&str>(const "explicit panic") -> [return: bb2, unwind: bb8];
     }
 
     bb2: {
