+++ tests/mir-opt/building/user_type_annotations.let_else.built.after.mir
@@ -35,7 +35,7 @@ fn let_else() -> () {
     }
 
     bb1: {
-        _1 = core::panicking::panic(const "internal error: entered unreachable code") -> bb6;
+        _1 = core::panicking::panic(const "internal error: entered unreachable code") -> [return: bb2, unwind: bb6];
     }
 
     bb2: {
